+++ CHANGES.md
@@ -1,3 +1,15 @@
+## v3.0.1 (2024-04-17)
+
+* Adress reviews from @reynir (#7, @samoht and @reynir)
+  - Fail when keys contain '.' and '..' to match other `mirage-kv-*`
+    implementations
+  - Use failwith instead of Lwt.failwith
+  - Use Lwt.reraise instead of Lwt.fail
+  - `digest` on directories now return an error
+  - `set_partial` on directories return an error while `set` on directories work
+    (and remove the directory)
+  - fix fd leak in `set` and `set_partial`
+
 ## v3.0.0 (2024-04-16)
 
 * Update to mirage-kv>6 (#5, @samoht)